Fix empty blame and file history dialogs #2822
The problem was in double `--` in the command line. The first one was added before file filter
if it's enabled, the second one was always added to the end of command line (#2700).

The problem reproduced only with `--follow` option in the command line
("Detect and follow renames"check box in Git Extensions).

The fix is to always add `--` before file filter (it doesn't matter if it's enabled or not).